requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backcall==0.2.0
cfgv==3.2.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
filelock==3.0.12
identify==2.1.2
importlib-metadata==3.7.2
importlib-resources==5.1.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
nodeenv==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.11.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.7.4.3
virtualenv==20.4.2
wcwidth==0.2.5
zipp==3.4.1
